+++ packages/dumb/src/index.ts
@@ -60,7 +60,7 @@ import { PrintView } from './modules/print/print-view/print-view'
 import { LayoutColumnType } from './models/enums/layout-column-type.enum'
 import { buildLayoutColumns } from './modules/print/print-view/print-view.util'
 import { getPatientFullName } from './utils/patient/patient.util'
-import { formatBirthdate, formatLocalizedFromUTC, getLongDayHourFormat, formatDate, TIMEZONE_UTC } from './utils/datetime/datetime.util'
+import { formatBirthdate, formatLocalizedFromUTC, getLongDayHourFormat, formatDate, TIMEZONE_UTC, formatCurrentDate } from './utils/datetime/datetime.util'
 import { renderPageNumbers } from './utils/pdf/pdf.util'
 import { WarmUpTooltip } from './components/tooltips/warm-up-tooltip/warm-up-tooltip'
 import { Device } from './models/device.model'
@@ -85,6 +85,7 @@ export {
   buildDevice,
   FocusedRangeLabels,
   FoodTooltip,
+  formatCurrentDate,
   formatBirthdate,
   getPatientFullName,
   AlarmEventTooltip,