+++ test/locales/translations.test.js
@@ -304,7 +304,6 @@ const ignoredTransKeyInYourLoopsFiles = [
   'IMEI',
   'Identifier',
   'Name',
-  'Parameters',
   'params|${parameter.name}',
   'Device',
   'Unit',
@@ -320,6 +319,7 @@ const ignoredTransKeyInYourLoopsFiles = [
   'Pump version',
   'Serial Number',
   'Setting',
+  'Settings on day',
   'Time In Range'
 ]