+++ CHANGELOG.md
@@ -11,6 +11,9 @@ It is based on Tidepool Blip 1.27.
 - YLP-3087 As a user, I want to see the medisafe pump event in yourloops
 - YLP-3141 As a user, I want my alarms messages to be the same as my device's
 
+### Fixed
+- YLP-2936 DBLG1 pdf settings are the ones of the current date
+
 ### Engineering use
 - YLP-2948 Updated failing tests because of the time change